third_party/blink/web_tests/platform/generic/fast/css/link-disabled-attr-parser-expected.txt

Adding the 'disabled' attribute to a link element causes it to disable its stylesheet

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(testDiv).backgroundColor is "rgb(255, 0, 0)"
PASS disabledLink.disabled is true
PASS successfullyParsed is true

TEST COMPLETE
